refactor: 예외처리시 원본 예외 함께 반환
