+++ jdbc/src/main/java/org/springframework/dao/DataUpdateException.java
@@ -1,8 +1,12 @@
-package org.springframework.jdbc.exception;
+package org.springframework.dao;
 
 public class DataUpdateException extends RuntimeException {
 
     public DataUpdateException(String message) {
         super(message);
+    }
+
+    public DataUpdateException(String message, Throwable cause) {
+        super(message, cause);
     }
 }
