--- jdbc/src/main/java/org/springframework/jdbc/exception/DataAccessException.java
@@ -1,8 +1,0 @@
-package org.springframework.jdbc.exception;
-
-public class DataAccessException extends RuntimeException{
-
-    public DataAccessException(String message) {
-        super(message);
-    }
-}
